# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DigiMontir Mobile API",
        "description": "Interactive OpenAPI documentation serving the technician endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API server instance"
        }
    ],
    "paths": {
        "/assignments": {
            "get": {
                "tags": [
                    "Assignments"
                ],
                "summary": "List all assignments with optional tracking parameters",
                "operationId": "5e676b85f8bcc603be0c177760a1b201",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "past",
                                "today",
                                "upcoming"
                            ]
                        }
                    },
                    {
                        "name": "status_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of Task objects"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/assignments/{id}/check-in": {
            "post": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Check In to an assignment location",
                "operationId": "9f810b6c4011c8eddc771a20845a4965",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "lat_check_in": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "lng_check_in": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "check_in_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-in registered successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/assignments/{id}/check-out": {
            "post": {
                "tags": [
                    "Assignments"
                ],
                "summary": "Check Out and finish the assignment",
                "operationId": "9073675a8c866cf57816e05d4f7f2e6d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "lat_check_out": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "lng_check_out": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "description_by_technician": {
                                        "type": "string"
                                    },
                                    "check_out_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-out saved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}